// src/enums.ts
export enum USER_ROLES_TYPES {
  SUPER_ADMIN = "SUPER_ADMIN",
  AGENCY_ADMIN = "AGENCY_ADMIN",
}

export enum AGENCY_TYPES {
  EQUAL_PARTS = "equalparts",
  LUMEN = "lumen",
  ASSURELY = "assurely",
}

export enum USER_STATUS {
  CONFIRMED = "CONFIRMED",
  PENDING = "PENDING",
  DISABLED = "DISABLED",
  FORCE_CHANGE_PASSWORD = "FORCE_CHANGE_PASSWORD",
} 